apps/fair-detr/requirements.txt
Brotli==1.0.7
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
dash==1.12.0
dash-core-components==1.10.0
dash-html-components==1.0.3
dash-renderer==1.4.1
dash-table==4.7.0
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
pandas==1.0.3
Pillow==7.1.2
plotly==4.8.0
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
retrying==1.3.3
scipy==1.4.1
six==1.15.0
torch==1.5.0
torchvision==0.6.0
urllib3==1.25.9
Werkzeug==1.0.1
